+++ tests/testthat/test_AutoTuner.R
@@ -219,11 +219,23 @@ test_that("store_tuning_instance, store_benchmark_result and store_models flags 
 
   assert_null(at$tuning_instance)
 
-  expect_error(AutoTuner$new(lrn("classif.rpart"), rsmp("holdout"), ms, te,
+  at = AutoTuner$new(lrn("classif.rpart"), rsmp("holdout"), ms, te,
+    tuner = tuner, ps, store_tuning_instance = FALSE, store_benchmark_result = FALSE,
+    store_models = TRUE)
+  at$train(task)
+
+  assert_r6(at$tuning_instance, "TuningInstanceSingleCrit")
+  assert_benchmark_result(at$tuning_instance$archive$benchmark_result)
+  assert_class(at$tuning_instance$archive$benchmark_result$resample_result(1)$learners[[1]]$model, "rpart")
+
+  at = AutoTuner$new(lrn("classif.rpart"), rsmp("holdout"), ms, te,
     tuner = tuner, ps, store_tuning_instance = FALSE, store_benchmark_result = TRUE,
-    store_models = FALSE),
-    regexp = "Benchmark results can only be stored if store_tuning_instance is set to TRUE",
-    fixed = TRUE)
+    store_models = FALSE)
+  at$train(task)
+
+  assert_r6(at$tuning_instance, "TuningInstanceSingleCrit")
+  assert_benchmark_result(at$tuning_instance$archive$benchmark_result)
+  assert_null(at$tuning_instance$archive$benchmark_result$resample_result(1)$learners[[1]]$model)
 })
 
 test_that("predict_type works", {
@@ -380,3 +392,60 @@ test_that("AutoTuner works with empty search space", {
   expect_list(at$tuning_instance$result$learner_param_vals[[1]], len = 0)
   expect_equal(at$tuning_instance$result$x_domain, list(list()))
 })
+
+test_that("AutoTuner importance method works", {
+  at = auto_tuner(
+    method = "random_search",
+    learner = lrn("classif.rpart"),
+    resampling = rsmp("cv", folds = 3),
+    measure = msr("classif.ce"),
+    term_evals = 4,
+    batch_size = 2
+  )
+
+  expect_error(at$importance(), "No model stored")
+  at$train(tsk("penguins"))
+  expect_numeric(at$importance(), len = 5)
+})
+
+test_that("AutoTuner selected_features method works", {
+  at = auto_tuner(
+    method = "random_search",
+    learner = lrn("classif.rpart"),
+    resampling = rsmp("cv", folds = 3),
+    measure = msr("classif.ce"),
+    term_evals = 4,
+    batch_size = 2
+  )
+
+  expect_error(at$selected_features(), "No model stored")
+  at$train(tsk("penguins"))
+  expect_character(at$selected_features())
+})
+
+test_that("AutoTuner oob_error method works", {
+  at = auto_tuner(
+    method = "random_search",
+    learner = lrn("classif.rpart"),
+    resampling = rsmp("cv", folds = 3),
+    measure = msr("classif.ce"),
+    term_evals = 4,
+    batch_size = 2
+  )
+
+  expect_error(at$oob_error(), "cannot calculate the out-of-bag error.")
+})
+
+test_that("AutoTuner loglik method works", {
+  at = auto_tuner(
+    method = "random_search",
+    learner = lrn("classif.rpart"),
+    resampling = rsmp("cv", folds = 3),
+    measure = msr("classif.ce"),
+    term_evals = 4,
+    batch_size = 2
+  )
+
+  expect_error(at$loglik(), "cannot calculate the log-likelihood.")
+})
+